(#20) Split screengrab.c
