+++ src/linux/screengrab.c
@@ -1,0 +1,35 @@
+#include "../screengrab.h"
+#include "../endian.h"
+#include <stdlib.h> /* malloc() */
+
+#include <X11/Xlib.h>
+#include <X11/Xutil.h>
+#include "../xdisplay.h"
+
+MMBitmapRef copyMMBitmapFromDisplayInRect(MMRect rect)
+{
+	MMBitmapRef bitmap;
+
+	Display *display = XOpenDisplay(NULL);
+	XImage *image = XGetImage(display,
+	                          XDefaultRootWindow(display),
+	                          (int)rect.origin.x,
+	                          (int)rect.origin.y,
+	                          (unsigned int)rect.size.width,
+	                          (unsigned int)rect.size.height,
+	                          AllPlanes, ZPixmap);
+	XCloseDisplay(display);
+	if (image == NULL) return NULL;
+
+	bitmap = createMMBitmap((uint8_t *)image->data,
+	                        rect.size.width,
+	                        rect.size.height,
+	                        (size_t)image->bytes_per_line,
+	                        (uint8_t)image->bits_per_pixel,
+	                        (uint8_t)image->bits_per_pixel / 8);
+	image->data = NULL; /* Steal ownership of bitmap data so we don't have to
+	                     * copy it. */
+	XDestroyImage(image);
+
+	return bitmap;
+}
