+++ src/macos/screengrab.c
@@ -1,0 +1,44 @@
+#include "../screengrab.h"
+#include "../endian.h"
+#include <stdlib.h> /* malloc() */
+
+#include <ApplicationServices/ApplicationServices.h>
+
+MMBitmapRef copyMMBitmapFromDisplayInRect(MMRect rect)
+{
+	MMBitmapRef bitmap = NULL;
+	uint8_t *buffer = NULL;
+	size_t bufferSize = 0;
+
+	CGDirectDisplayID displayID = CGMainDisplayID();
+
+	CGImageRef image = CGDisplayCreateImageForRect(displayID,
+		CGRectMake(rect.origin.x,
+			rect.origin.y,
+			rect.size.width,
+			rect.size.height));
+
+	if (!image) { return NULL; }
+
+	CFDataRef imageData = CGDataProviderCopyData(CGImageGetDataProvider(image));
+
+	if (!imageData) { return NULL; }
+
+	bufferSize = CFDataGetLength(imageData);
+	buffer = malloc(bufferSize);
+
+	CFDataGetBytes(imageData, CFRangeMake(0,bufferSize), buffer);
+
+	bitmap = createMMBitmap(buffer,
+		CGImageGetWidth(image),
+		CGImageGetHeight(image),
+		CGImageGetBytesPerRow(image),
+		CGImageGetBitsPerPixel(image),
+		CGImageGetBitsPerPixel(image) / 8);
+
+	CFRelease(imageData);
+
+	CGImageRelease(image);
+
+	return bitmap;
+}
