+++ src/win32/screengrab.c
@@ -1,0 +1,71 @@
+#include "../screengrab.h"
+#include "../endian.h"
+#include <stdlib.h> /* malloc() */
+
+MMBitmapRef copyMMBitmapFromDisplayInRect(MMRect rect)
+{
+	MMBitmapRef bitmap;
+	void *data;
+	HDC screen = NULL, screenMem = NULL;
+	HBITMAP dib;
+	BITMAPINFO bi;
+
+	/* Initialize bitmap info. */
+	bi.bmiHeader.biSize = sizeof(bi.bmiHeader);
+   	bi.bmiHeader.biWidth = (long)rect.size.width;
+   	bi.bmiHeader.biHeight = -(long)rect.size.height; /* Non-cartesian, please */
+   	bi.bmiHeader.biPlanes = 1;
+   	bi.bmiHeader.biBitCount = 32;
+   	bi.bmiHeader.biCompression = BI_RGB;
+   	bi.bmiHeader.biSizeImage = (DWORD)(4 * rect.size.width * rect.size.height);
+	bi.bmiHeader.biXPelsPerMeter = 0;
+	bi.bmiHeader.biYPelsPerMeter = 0;
+	bi.bmiHeader.biClrUsed = 0;
+	bi.bmiHeader.biClrImportant = 0;
+
+	screen = GetDC(NULL); /* Get entire screen */
+	if (screen == NULL) return NULL;
+
+	/* Get screen data in display device context. */
+   	dib = CreateDIBSection(screen, &bi, DIB_RGB_COLORS, &data, NULL, 0);
+
+	/* Copy the data into a bitmap struct. */
+	if ((screenMem = CreateCompatibleDC(screen)) == NULL ||
+	    SelectObject(screenMem, dib) == NULL ||
+	    !BitBlt(screenMem,
+	            (int)0,
+	            (int)0,
+	            (int)rect.size.width,
+	            (int)rect.size.height,
+				screen,
+				rect.origin.x,
+				rect.origin.y,
+				SRCCOPY)) {
+		
+		/* Error copying data. */
+		ReleaseDC(NULL, screen);
+		DeleteObject(dib);
+		if (screenMem != NULL) DeleteDC(screenMem);
+
+		return NULL;
+	}
+
+	bitmap = createMMBitmap(NULL,
+	                        rect.size.width,
+	                        rect.size.height,
+	                        4 * rect.size.width,
+	                        (uint8_t)bi.bmiHeader.biBitCount,
+	                        4);
+
+	/* Copy the data to our pixel buffer. */
+	if (bitmap != NULL) {
+		bitmap->imageBuffer = malloc(bitmap->bytewidth * bitmap->height);
+		memcpy(bitmap->imageBuffer, data, bitmap->bytewidth * bitmap->height);
+	}
+
+	ReleaseDC(NULL, screen);
+	DeleteObject(dib);
+	DeleteDC(screenMem);
+
+	return bitmap;
+}
